+++ tests/ApiCodeBaseTest.php
@@ -2,7 +2,7 @@
 
 namespace MarcinOrlowski\ResponseBuilder\Tests;
 
-use MarcinOrlowski\ResponseBuilder\ApiCodeBase;
+use MarcinOrlowski\ResponseBuilder\BaseApiCodes;
 
 /**
  * Laravel API Response Builder
@@ -14,7 +14,7 @@ use MarcinOrlowski\ResponseBuilder\ApiCodeBase;
  * @license   http://www.opensource.org/licenses/mit-license.php MIT
  * @link      https://github.com/MarcinOrlowski/laravel-api-response-builder
  */
-class ApiCodeBaseTest extends Base\ResponseBuilderTestCaseBase
+class ApiCodeBaseTest extends TestCase
 {
 	/**
 	 * Tests getMinCode() with invalid config
@@ -26,7 +26,7 @@ class ApiCodeBaseTest extends Base\ResponseBuilderTestCaseBase
 	public function testGetMinCode_MissingConfigKey()
 	{
 		\Config::offsetUnset('response_builder.min_code');
-		ApiCodeBase::getMinCode();
+		BaseApiCodes::getMinCode();
 	}
 
 	/**
@@ -39,7 +39,7 @@ class ApiCodeBaseTest extends Base\ResponseBuilderTestCaseBase
 	public function testGetMaxCode_MissingConfigKey()
 	{
 		\Config::offsetUnset('response_builder.max_code');
-		ApiCodeBase::getMaxCode();
+		BaseApiCodes::getMaxCode();
 	}
 
 
@@ -53,7 +53,7 @@ class ApiCodeBaseTest extends Base\ResponseBuilderTestCaseBase
 	public function testGetMap_MissingConfigKey()
 	{
 		\Config::offsetUnset('response_builder.map');
-		ApiCodeBase::getMap();
+		BaseApiCodes::getMap();
 	}
 
 	/**
@@ -66,7 +66,7 @@ class ApiCodeBaseTest extends Base\ResponseBuilderTestCaseBase
 	public function testGetMap_WrongConfig()
 	{
 		\Config::set('response_builder.map', false);
-		ApiCodeBase::getMap();
+		BaseApiCodes::getMap();
 	}
 
 
@@ -79,7 +79,7 @@ class ApiCodeBaseTest extends Base\ResponseBuilderTestCaseBase
 	 */
 	public function testGetCodeMessageKey_OutOfRange()
 	{
-		ApiCodeBase::getCodeMessageKey(ApiCodeBase::RESERVED_MAX_API_CODE + 1);
+		BaseApiCodes::getCodeMessageKey(BaseApiCodes::RESERVED_MAX_API_CODE + 1);
 	}
 
 	/**
@@ -91,7 +91,7 @@ class ApiCodeBaseTest extends Base\ResponseBuilderTestCaseBase
 	 */
 	public function testGetReservedCodeMessageKey_TooLow()
 	{
-		ApiCodeBase::getReservedCodeMessageKey(ApiCodeBase::RESERVED_MIN_API_CODE - 1);
+		BaseApiCodes::getReservedCodeMessageKey(BaseApiCodes::RESERVED_MIN_API_CODE - 1);
 	}
 
 	/**
@@ -103,7 +103,7 @@ class ApiCodeBaseTest extends Base\ResponseBuilderTestCaseBase
 	 */
 	public function testGetReservedCodeMessageKey_TooHigh()
 	{
-		ApiCodeBase::getReservedCodeMessageKey(ApiCodeBase::RESERVED_MAX_API_CODE + 1);
+		BaseApiCodes::getReservedCodeMessageKey(BaseApiCodes::RESERVED_MAX_API_CODE + 1);
 	}
 
 	/**
@@ -114,19 +114,19 @@ class ApiCodeBaseTest extends Base\ResponseBuilderTestCaseBase
 	public function testGetReservedCodeMessageKey()
 	{
 		// check how mapped code handling works
-		$mapping = ApiCodeBase::getReservedCodeMessageKey(ApiCodeBase::OK);
+		$mapping = BaseApiCodes::getReservedCodeMessageKey(BaseApiCodes::OK);
 		$this->assertNotNull($mapping);
 
 		// check how not-mapped code is handled
-		$base_map = $this->getProtectedMember(ApiCodeBase::class, 'base_map');
+		$base_map = $this->getProtectedMember(BaseApiCodes::class, 'base_map');
 
-		for ($code = ApiCodeBase::RESERVED_MIN_API_CODE; $code < ApiCodeBase::RESERVED_MAX_API_CODE; $code++) {
+		for ($code = BaseApiCodes::RESERVED_MIN_API_CODE; $code < BaseApiCodes::RESERVED_MAX_API_CODE; $code++) {
 			if (!array_key_exists($code, $base_map)) {
 				break;
 			}
 		}
 
-		$mapping = ApiCodeBase::getReservedCodeMessageKey($code);
+		$mapping = BaseApiCodes::getReservedCodeMessageKey($code);
 		$this->assertNull($mapping);
 	}