+++ tests/ErrorTest.php
@@ -15,9 +15,8 @@ use Symfony\Component\HttpFoundation\Response as HttpResponse;
  * @license   http://www.opensource.org/licenses/mit-license.php MIT
  * @link      https://github.com/MarcinOrlowski/laravel-api-response-builder
  */
-class ErrorTest extends Base\ResponseBuilderTestCaseBase
+class ErrorTest extends TestCase
 {
-
 	/**
 	 * Check success()
 	 *
@@ -197,7 +196,7 @@ class ErrorTest extends Base\ResponseBuilderTestCaseBase
 	 */
 	public function testError_MissingMessageMapping()
 	{
-		/** @var \MarcinOrlowski\ResponseBuilder\ApiCodeBase $api_codes_class_name */
+		/** @var \MarcinOrlowski\ResponseBuilder\BaseApiCodes $api_codes_class_name */
 		$api_codes_class_name = $this->getApiCodesClassName();
 
 		// FIXME we **assume** this is not mapped. But assumptions sucks...
@@ -221,7 +220,7 @@ class ErrorTest extends Base\ResponseBuilderTestCaseBase
 	 */
 	public function testBuildErrorResponse_ApiCodeOK()
 	{
-		/** @var \MarcinOrlowski\ResponseBuilder\ApiCodeBase $api_codes_class_name */
+		/** @var \MarcinOrlowski\ResponseBuilder\BaseApiCodes $api_codes_class_name */
 		$api_codes_class_name = $this->getApiCodesClassName();
 
 		$data = null;
@@ -259,7 +258,7 @@ class ErrorTest extends Base\ResponseBuilderTestCaseBase
 	 */
 	public function testBuildErrorResponse_WrongHttpCodeType()
 	{
-		/** @var \MarcinOrlowski\ResponseBuilder\ApiCodeBase $api_codes_class_name */
+		/** @var \MarcinOrlowski\ResponseBuilder\BaseApiCodes $api_codes_class_name */
 		$api_codes_class_name = $this->getApiCodesClassName();
 
 		$data = null;
@@ -277,7 +276,7 @@ class ErrorTest extends Base\ResponseBuilderTestCaseBase
 	 */
 	public function testBuildErrorResponse_NullHttpCode()
 	{
-		/** @var \MarcinOrlowski\ResponseBuilder\ApiCodeBase $api_codes_class_name */
+		/** @var \MarcinOrlowski\ResponseBuilder\BaseApiCodes $api_codes_class_name */
 		$api_codes_class_name = $this->getApiCodesClassName();
 
 		$data = null;
@@ -300,7 +299,7 @@ class ErrorTest extends Base\ResponseBuilderTestCaseBase
 	 */
 	public function testBuildErrorResponse_TooLowHttpCode()
 	{
-		/** @var \MarcinOrlowski\ResponseBuilder\ApiCodeBase $api_codes_class_name */
+		/** @var \MarcinOrlowski\ResponseBuilder\BaseApiCodes $api_codes_class_name */
 		$api_codes_class_name = $this->getApiCodesClassName();
 
 		$data = null;
@@ -320,7 +319,7 @@ class ErrorTest extends Base\ResponseBuilderTestCaseBase
 	 */
 	public function testBuildErrorResponse_WrongLangArgs()
 	{
-		/** @var \MarcinOrlowski\ResponseBuilder\ApiCodeBase $api_codes_class_name */
+		/** @var \MarcinOrlowski\ResponseBuilder\BaseApiCodes $api_codes_class_name */
 		$api_codes_class_name = $this->getApiCodesClassName();
 
 		$data = null;